% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_genbank.R
\name{download_genbank}
\alias{download_genbank}
\title{download_genbank}
\usage{
download_genbank(acc, db, destfile)
}
\arguments{
\item{acc}{The accession number}

\item{db}{The database, it can be nuccore, nucest, nucgss or popset}

\item{destfile}{The path where you want to save the genbank files}
}
\description{
This function downloads genbank files from NCBI.
}
