% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_align.R
\name{global_aln}
\alias{global_aln}
\title{global_aln}
\usage{
global_aln(X, Y, Scoring_system)
}
\arguments{
\item{X}{A sequence to be aligned}

\item{Y}{A sequence to be aligned}

\item{Scoring_system}{A vector containing the score of match, mismatch and
inserting a gap}
}
\description{
This function reads the fasta files, extract the sequence from
    it and calculate the proportion of each kind of base in sequence.
}
