% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.AlignmentResult.R
\name{print.AlignmentResult}
\alias{print.AlignmentResult}
\title{print.AlignmentResult}
\usage{
\method{print}{AlignmentResult}(x, ...)
}
\arguments{
\item{x}{The object returned by function global_align}

\item{...}{Other parameter}
}
\description{
This function prints the object returned by function
global_align.
}
