% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.readfasta.R
\name{print.readfasta}
\alias{print.readfasta}
\title{print.readfasta}
\usage{
\method{print}{readfasta}(x, ...)
}
\arguments{
\item{x}{The object returned by function read.fasta}

\item{...}{Other parameter pass to print}
}
\description{
This function prints the object returned by function read.fasta.
}
