% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fasta.R
\name{read.fasta}
\alias{read.fasta}
\title{read.fasta}
\usage{
read.fasta(fa_files)
}
\arguments{
\item{fa_files}{A character vector in which all elements are the
absolute path of fasta files.}
}
\description{
This function reads the fasta files, extract the sequence from
it and calculate the proportion of each kind of base in sequence.
}
