% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.genbank.R
\name{read.genbank}
\alias{read.genbank}
\title{read.genbank}
\usage{
read.genbank(gb_files)
}
\arguments{
\item{gb_files}{A character vector in which all elements are the
absolute path of genbank files}
}
\description{
This function reads the genbank files, extracts the sequence
from it, calculates the proportion of each kind of base in sequence
and generates a fasta file which contains the sequence in it.
}
